+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### To be Released
 
+* Bugfix: integration-link-create: command will always fail if there's already a link [#520](https://github.com/Scalingo/cli/issues/520)
+
 ### 1.16.5
 
 * Add `--bind` arg to `db-tunnel` command that let us bind a custom host (and not only 127.0.0.1) [#517](https://github.com/Scalingo/cli/pull/517)